requirements.txt
discord
bson
pillow
pymongo